Fix external function addresses in linker script
The reason for this is bizarre. ARM treats branch targets interestingly,
since addresses must be aligned on at least 2-byte boundaries. The LSB,
when set, indicates that the target instructions use the Thumb encoding.

The C compiler and binutils used in the version of the compiler
downloaded by the scripts in this repo did what's probably the correct
thing and added 1 to every constant address listed in the linker symbol
file when encountering a call to such a symbol. The version of the
linker included with the Rust compiler for the thumbv7em-none-eabihf
target doesn't seem to do this, and therefore generated code that
crashed. Adding 1 to every function address in the linker script seems
to fix this.